requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.6
redis>=4.5.5
celery>=5.3.1
django-celery-beat>=2.5.0
python-dotenv>=1.0.0
drf-yasg>=1.21.7
gunicorn>=21.2.0
django-environ>=0.10.0
django-cors-headers>=4.1.0
django-redis>=5.3.0
